Half-full glass of Matrix classes and serialization

+++ composeApp/src/commonMain/kotlin/data/io/social/network/conversation/message/MediaIO.kt
@@ -17,7 +17,7 @@ data class MediaIO(
     val name: String? = null,
 
     /** Size in bytes of the media */
-    val size: Int? = null,
+    val size: Long? = null,
 
     /** Local file path */
     @Ignore
